+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.1.7 -- added promises
+
+1. Added promise support (why did I not do this before..)
+
+# 0.1.6 -- fix RPC error handling
+
+1. fix RPC error handling for geth nodes
+
+# 0.1.5 -- package json fix
+
+# 0.1.4 -- handle 405
+
+1. Handle 405 errors better form the provider
+
 # 0.1.2 -- louder errors
 
 1. Louder errors